+++ app/webapp/settings.py
@@ -54,6 +54,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.sites",
+    "django.contrib.humanize",
     "django_htmx",
     "guardian",
     "allauth",
