--- app/webapp/templates/webapp/vessels/note_message_add.html
@@ -1,25 +1,0 @@
-{% extends "webapp/base.html" %}
-
-{% block title %}Add Message to Note - {{ vessel.name }}{% endblock %}
-
-{% block content %}
-<div class="max-w-2xl mx-auto mt-10">
-    <div class="bg-white rounded-lg shadow p-6">
-        <h1 class="text-2xl font-bold mb-4">
-            {% if note.owner == user %}
-                Note for {{ vessel.name }}{% if note.shared_with.all %} with {% for u in note.shared_with.all %}{{ u.username }}{% if not forloop.last %}, {% endif %}{% endfor %}{% endif %}
-            {% else %}
-                Note for {{ vessel.name }} with {{ note.owner.username }}
-            {% endif %}
-        </h1>
-        <form method="post" id="note-message-form">
-            {% csrf_token %}
-            {% include "webapp/components/milkdown_editor.html" with field_name="content" field_id="note-message-content" label="Message" initial_value="" errors=form.content.errors form_id="note-message-form" %}
-            <div class="flex justify-end">
-                <a href="{% url 'vessel_detail' vessel.pk %}" class="mr-4 px-4 py-2 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50">Cancel</a>
-                <button type="submit" class="px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-accent hover:bg-primary">Add Message</button>
-            </div>
-        </form>
-    </div>
-</div>
-{% endblock %}
